Analytics

+++ Dentizone.Application/DTOs/Analytics/PostAnalyticsDTO.cs
@@ -10,7 +10,7 @@ namespace Dentizone.Application.DTOs.Analytics
     {
         public int TotalPosts { get; set; }
 
-        public int AveragePostPrice { get; set; }
+        public decimal AveragePostPrice { get; set; }
         public Dictionary<string, int> PostsByCategory { get; set; } = new Dictionary<string, int>();
 
     }
